Move file copying helper methods to the TestHelpers project
